Replace `const string&` with `string_view` where possible
`string_view` is technically nicer and faster, but as it turns out,
rarely useable: there is no implicit conversion to string, if a
function argument requires a string and a `string_view` is passed; a
C-style string, "text", can't be passed to a function if it accepts a
string_view; and `fstream.open()` does not allow for a `string_view`
to be passed.

So while a few functions have been `string_view`-ified, most functions
accept a `const string&`, since that works better overall.