Make type conversions explicit
This is somewhat pedantic, but type conversion have caused problems in
the past. Making them explicit hopefully shows clearer where such
potential problems are.

Nearly all type conversions are from size_t to double; as long as
size_t is reasonably small (which it is), the precision will be
conserved.

A few variables and a getter function needed to get their type changed
to size_t as well.