Remove size member variable
The `size` member variable is unnecessary, since the vector member
variables keep their size with them. The only potential issue is that
this doesn't guarantee that multiple member vectors have the same
size. This could be solved by using a vector of a struct, but access
is likely to be (a lot) slower when iterating through such a struct
vector.